+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 #### [Unreleased]
 
+## Updated
+
+- Fixed the Active.Trigger elements failing to render anchor nodes when passed
+  an href.
+
 ## [3.0.0-alpha.8] - 2018-10-19
 
 ## BREAKING CHANGES